[PATCH] Relocation overflow with modules on Alpha
From: Ivan Kokshaysky <ink@jurassic.park.msu.ru>

Larger modules fail to load with the message "Relocation overflow vs
section 17", or some other section number.

This failure happens with GPRELHIGH relocation, which is *signed* short,
but relocation overflow check in module.c doesn't take into account the
sign extension.

+++ arch/alpha/kernel/module.c
@@ -259,7 +259,7 @@ apply_relocate_add(Elf64_Shdr *sechdrs, const char *strtab,
 			*(u64 *)location = value;
 			break;
 		case R_ALPHA_GPRELHIGH:
-			value = (value - gp + 0x8000) >> 16;
+			value = (long)(value - gp + 0x8000) >> 16;
 			if ((short) value != value)
 				goto reloc_overflow;
 			*(u16 *)location = value;
